requirements.txt
flask==3.0.0
python-dotenv
langchain
openai